Split network into ContentMesh modules
refactor(network): rename MeshClient to ContentMesh; split order,
 packaging, named_binds, bootstrap, and node_http;
feat(network): add feedback/identity seams and CoDTransport stub;
refactor(network): move legacy CoD under network/legacy;
 update package entrypoints to ContentMesh
