+++ cats/__init__.py
@@ -1,18 +1,18 @@
 import os
 from os.path import dirname, abspath
 
-from cats.network import MeshClient
+from cats.network import ContentMesh
 from cats.network.clients.ipfs_client import connect as connect_ipfs
 from cats.service import Service
 
 CWD = os.getcwd()
 CATS_HOME = dirname(dirname(abspath(__file__)))
-MESH_CLIENT = MeshClient(
+CONTENT_MESH = ContentMesh(
     ipfsClient=connect_ipfs(),
     CATS_HOME=CATS_HOME,
 )
 SERVICE = Service(
-    meshClient=MESH_CLIENT,
+    contentMesh=CONTENT_MESH,
     CATS_HOME=CATS_HOME
 )
 DATA_HOME = SERVICE.DATA_HOME
