+++ cats/network/feedback/__init__.py
@@ -1,0 +1,4 @@
+"""Unsigned Invoice/BOM feedback packaging (PROV/VC later)."""
+from cats.network.feedback.envelope import attach_node_uri, build_execution_bom
+
+__all__ = ['attach_node_uri', 'build_execution_bom']
