+++ cats/network/identity/__init__.py
@@ -1,0 +1,4 @@
+"""CAT Node identity seams (node URI string today; DID later)."""
+from cats.network.identity.node_uri import node_uri
+
+__all__ = ['node_uri']
