+++ cats/network/legacy/__init__.py
@@ -1,0 +1,4 @@
+"""Quarantined network helpers (not ContentMesh / CoDTransport public API)."""
+from cats.network.legacy.cod import CoD
+
+__all__ = ["CoD"]
